backend/authentication_service/requirements.txt
asyncio
nats-py
bcrypt
pyjwt